Add webppl version of polite RSA
To run in command line, `webppl politeness.wppl --require utils`

This model is very similar to the mixture of goals
politeness.church model. Worlds are now
`{state, valence}` objects (rather than have
a deterministic function from state-> valence).

Pragmatic listener has uncertainty over the speaker's goal.
Mixture of goals is permitted at the pragmatic listener level.
Speaker1 gets passed a world---{state, valence}--- and a
weight vector of goals. Speaker1 samples a goal, and then chooses
an utterance in accord with that goal. If goal is honesty, then
communicate the state. If goal is not honesty, then communicate
valence. Valence is not the valence that corresponds to the state,
however. Valence is sampled (via `beingNiceOrMean`) with increasing
probability assigned to higher valence values (if kind) or
decreasing probability assigned to lower valence values (if not kind).
This, in conjunction with `qudFunction` (which in this case, will return
a function that projects onto valence only), effectively misrepresents
the true state.