New nice production function. Add rounded data.
Nice production used to be state-invariant. Now, it takes state
as an argument, and select states with valence greater than or
equal to the true state, with increasing probabilities. States for
which the valence is less than the true state, are assigned very low
production probabilities. This, I believe, fixed an issue that arose
when the true state was AMAZING and speaker said "good". With a production
function invariant of state, "good" was still relatively nice (as opposed to mean)
but it was honest, and so the speaker was presumed to be extremely nice. This
violated intuitions.